ContextStore moved to context.py
