+++ src/main.py
@@ -9,7 +9,7 @@ from constants import CSV, JSONLD
 from custom_logging import logger
 from middleware import logging_middleware
 from store import StubCsvStore, StubMetadataStore
-from store.metadata.base import ContextStore
+from store.metadata.context import ContextStore
 
 # Simple env var flag to allow local browsing of api responses
 # while developing.
