+++ src/store/metadata/base.py
@@ -88,12 +88,3 @@ class BaseMetadataStore(ABC):
         """
         Get a specific sub-topic for a specific topic
         """
-
-
-class ContextStore(ABC):
-    def __init__(self):
-        with open(Path("src/store/metadata/context.json").absolute(), "r") as json_file:
-            self.context = json.load(json_file)
-
-    def get_context(self):
-        return self.context
